create  category, delete category, set and update category on meals

+++ database/migrations/2020_02_18_154627_create_side_to_meal_table.php
@@ -15,8 +15,8 @@ class CreateSideToMealTable extends Migration
     {
         Schema::create('side_to_meal', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('mealid');
-            $table->integer('sideid');
+            $table->bigInteger('mealid');
+            $table->bigInteger('sideid');
             $table->timestamps();
         });
     }
